Extract browser runner boot phase templates
